pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "medextract"
version = "0.1.0"
description = "Extract readings from medical device images using a local LLM (Ollama + MedGemma)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["medical", "blood-pressure", "ocr", "llm", "ollama"]
dependencies = [
    "ollama>=0.6.1",
    "pillow>=10.2.0",
    "pandas>=2.1.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-mock>=3.12"]

[project.scripts]
medextract = "medextract.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["medextract*"]
